CI: deploy for all branches
When you tag a commit on master and push it to GitHub, TravisCI will actually
run two builds, one for master and one for a branch named after the tag.

We use a conditional deploy stage. It will be only used if the git tag happens
to match a certain regexp.

By default the deployment provider only deploys for the master branch. That will
never be the case for our conditional stage. So we deploy for all branches now,
knowing it will only happen for tag branches anyway.