Support $NVIM
Latest Neovim from the repo does a few changes to how $NVIM_LISTEN_ADDRESS is
handled. You can still use it to set a servername, but it's not passed on to
children of the process. That also affects the use of :terminal, which will now
have a $NVIM set instead. $NVIM points to the server exactly like
$NVIM_LISTEN_ADDRESS did.

So all that changes, is that we now also check for $NVIM before
$NVIM_LISTEN_ADDRESS.

References https://github.com/neovim/neovim/pull/11009